Move all MIPS decoding to a separate translation unit
