+++ simulator/mips/mips_instr.h
@@ -11,7 +11,6 @@
 #include "mips_register/mips_register.h"
 #include "mips_version.h"
 
-// MIPT-MIPS modules
 #include <func_sim/trap_types.h>
 #include <infra/endian.h>
 #include <infra/exception.h>
@@ -19,10 +18,6 @@
 #include <infra/string_view.h>
 #include <infra/types.h>
 #include <kryucow_string.h>
-
-// Generic C++
-#include <array>
-#include <unordered_map>
 
 struct MIPSInstrDecoder;
